Allow finding `spicyz` in `PATH`.
If `spicyz` was not installed through `zkg` we would previously not have
found it. With this patch we can find it in the environment, or if that
fails through in `zkg`'s store. This allows us to build both against an
older zkg which did not inject package bin directories into its `PATH`,
and also against a spicy-plugin bundled with Zeek as is the default with
zeek-5.