Fix string representation of extended search filters
We want to render search filters similar to e.g., Wireshark. In order to
calculate a string representation of extended search filter we need to
output their values in a order which is different from other fields.

This patch fixes this directly, but the bigger issue is probably that
`DecodedAttributeValue` is a too generic unit and we should have
lower-level types for constructs like extended search filters which know
how to render themselves correctly.

Closes #23.